% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{achievements}
\alias{achievements}
\title{Animals Dataset}
\format{
A data frame with 30 rows and 6 variables:
\describe{
  \item{id}{the achievement id}
  \item{name}{the name of the achievement}
  \item{description}{a longer description of the achievement}
  \item{display_on_collections_tab_before_earned}{?}
  \item{prerequisite_achievement}{achievement that must be completed first}
  \item{hat_earned}{the id of the hat earned by the achievement, which can
  be joined to the column `object_id` in the `hats` dataset}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Hat_data}
}
\usage{
achievements
}
\description{
A dataset containing information about achievements in Stardew Valley.
Achievements are in-game events or tasks which, when completed, make a new
hat available for purchase in the hat shop.
}
\keyword{datasets}
