% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_blueprints}
\alias{animal_blueprints}
\title{Animal Blueprints}
\format{
A data frame with 7 rows and 7 variables:
\describe{
  \item{animal_name}{animal name}
  \item{is_animal}{always 'animal', and only included here in case of future game
  updates}
  \item{purchase_price}{price to purchase the animal}
  \item{source_rectangle_for_menu_view_x}{x-coordinate of the menu
  rectangle}
  \item{source_rectangle_for_menu_view_y}{y-coordinate of the menu
  rectangle}
  \item{description}{description of the animal}
  \item{is_farm}{currently always 'Farm' and only included here in case of
  future game updates}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Animal_data}
}
\usage{
animal_blueprints
}
\description{
A dataset containing "blueprints" for purchasing animals.
}
\keyword{datasets}
