% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{building_blueprints}
\alias{building_blueprints}
\title{Building Blueprints Data}
\format{
A data frame with 51 rows and 21 variables:
\describe{
  \item{building_name}{the name of the building}
  \item{required_object_id}{the object id of the item required to build
  the building}
  \item{object_name}{the name of the object required to build the building}
  \item{required_quantity}{the number of the object required for
  making the building}
  \item{tiles_width}{width of the building in game tiles}
  \item{tiles_height}{height of the building in game tiles}
  \item{human_door_x_coord}{x-coordinate of the door that players enter
  through}
  \item{human_door_y_coord}{y-coordinate of the door that players enter
  through}
  \item{animal_door_x_coord}{x-coordinate of the door that animals enter
  through}
  \item{animal_door_y_coord}{y-coordinate of the door that animals enter
  through}
  \item{map_to_warp_to}{the map image the game uses for the interior of
  the building, if any}
  \item{description}{a description of the building}
  \item{blueprint_type}{whether a building is an initial building or
  an upgraded version of a previous building}
  \item{name_of_building_to_upgrade}{if building type is 'upgrade', then the
  name of the building to upgrade}
  \item{source_rectangle_for_menu_view_x_coord}{x-coordinate of the menu view
  rectangle}
  \item{source_rectangle_for_menu_view_y_coord}{y-coordinate of the menu view
  rectangle}
  \item{max_occupants}{maximum number of occupants}
  \item{action_behavior}{action the building provides, if any}
  \item{names_of_building_location}{where the building is located}
  \item{price}{cost of having the building constructed}
  \item{magical}{is the building magical?}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game.
Descriptions of the variables summarized from \url{https://stardewvalleywiki.com/Modding:Big_craftables_data}
}
\usage{
building_blueprints
}
\description{
A dataset containing information about blueprints for buildings you can
purchase or construct.
}
\keyword{datasets}
