% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundles}
\alias{bundles}
\title{Bundles Data}
\format{
A data frame with 31 rows and 10 variables:
\describe{
  \item{room_id}{Name of the room which combined with sprite_index will
  uniquely identify each row (each room has multiple bundles)}
  \item{sprite_index}{The index of the sprite used for the bundle image}
  \item{bundle_name}{the name of the bundle}
  \item{reward_type_of_object}{The type of the reward earned by delivering
  the bundle, where the type may be O for "object," BO for "big object,"
  F for "furniture," H  for "hat", C for "clothing", or R for "ring". A bundle
  reward may include more than one type.}
  \item{reward_object_id}{id of the object given as a reward
  for completing the bundle}
  \item{reward_number_of_object_given}{The count of the object given as a reward}
  \item{required_objects}{Space-delimited triples for the items required in
  the bundle. Each triple contains (in order) the object id, the number
  of objects required, and the minimum quality of object. If the first value
  (for object id) is -1, then the required object is an amount of gold, and
  both the quality and number of items required should be the same,
  representing the amount of gold.}
  \item{required_object_is_gold}{Type logical representing if the object required
  to complete the quest is gold}
  \item{color_index}{Color index of the bundle color}
  \item{min_number_of_items}{a description of the building}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game.
Descriptions of the variables summarized from \url{https://stardewvalleywiki.com/Modding:Bundles}
}
\usage{
bundles
}
\description{
A dataset containing information about the bundles delivered to the Community
Center.
}
\keyword{datasets}
