% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seed_prices.R
\name{calc_seed_prices}
\alias{calc_seed_prices}
\title{Calc Seed Prices}
\usage{
calc_seed_prices(all_joja_cols = TRUE, joja_member = FALSE)
}
\arguments{
\item{joja_member}{logical specifying if you have the JojaMart membership,
or can be NA if both_joja_cols = FALSE}

\item{crop_data}{data.frame containing crop data, defaults to the data
contained in the croptimizer package}

\item{both_joja_cols}{logical specifying if you want both JojaMart price
columns, one for the price with the membership and one for the price without}
}
\value{
data.frame containing crop_data with seed_price and source columns
appended
}
\description{
Calculates the seed prices for purchased seeds based upon if you have a
Jojamart membership or not.
}
\examples{
fall_crops <-
filter_to_season(season = "fall", crop_data = croptimizer::crops)
calc_seed_prices(crop_data = fall_crops,
joja_member = FALSE,
drop_cols = TRUE)
}
