% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caught_fish}
\alias{caught_fish}
\title{Caught Fish}
\format{
A data frame with 55 rows and 15 variables:
\describe{
  \item{object_id}{The object id of the fish}
  \item{name}{The name of the fish}
  \item{chance_of_darting}{The probability of the fish making a dart}
  \item{darting_randomness}{A measure of the degree of randomness in darting}
  \item{min_size}{Minimum size of the fish}
  \item{max_size}{Maximum size of the fish}
  \item{min_time_of_spawn}{Earliest time of day for the spawn}
  \item{max_time_of_spawn}{Latest time of day for the spawn}
  \item{weather}{Weather during which the fish can appear}
  \item{locations}{Locations where the fish can appear}
  \item{min_depth}{Minimum depth where the fish can appear, where depth is
  the distance the bobber is from the shore}
  \item{spawn_multiplier}{The multiplier used for spawn rate calculations}
  \item{depth_multiplier}{The depth multiplier use in spawn rate calculations}
  \item{min_fishing_level}{The minimum character fishing level required for this
  fish to appear}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game.
Descriptions of the variables summarized from \url{https://stardewvalleywiki.com/Modding:Fish_data}
}
\usage{
caught_fish
}
\description{
A dataset containing information about fish that are caught with a fishing pole.
}
\details{
The spawn rate for fish functions by first shuffling the fish that are eligible
based on season, location, and time of day (plus seaweed/algae), and then calculating the base
spawn probability for the first fish of that list:

 base spawn probability = spawn multiplier - max(0, minimum depth - actual depth) × depth multiplier × spawn multiplier + fishing level / 50

 This probability is capped at a maximum of 90%. Then, the game rolls to determine
 if that fish will successfully spawn. If it does, then the player has a chance
 to catch it. If it does not succeed, then the next fish in the list is rolled
 for. If no fish succeed, then a random trash item spawns.
}
\keyword{datasets}
