% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complete_npc_gift_tastes}
\alias{complete_npc_gift_tastes}
\title{Complete NPC Gift Tastes}
\format{
A data frame with 33 rows and 11 variables:
\describe{
  \item{character_name}{the NPC name}
  \item{love_dialogue}{dialogue NPC uses when given a loved item}
  \item{loved_items}{object ID's of the loved items}
  \item{liked_dialogue}{dialogue NPC uses when given a liked item}
  \item{liked_items}{object ID's of the liked items}
  \item{dislike_dialogue}{dialogue NPC uses when given a disliked item}
  \item{disliked_items}{object ID's of the disliked items}
  \item{hate_dialogue}{dialogue NPC uses when given a hated item}
  \item{hated_items}{object ID's of the hated items}
  \item{neutral_dialogue}{dialogue NPC uses when given a neutral item}
  \item{neutral_items}{object ID's of the neutral items}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Gift_taste_data}
}
\usage{
complete_npc_gift_tastes
}
\description{
A dataset containing NPC gift tastes, including universal tastes.
}
\keyword{datasets}
