% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_na.R
\name{convert_to_na}
\alias{convert_to_na}
\title{Convert to NA}
\usage{
convert_to_na(x, value)
}
\arguments{
\item{x}{a vector to recode}

\item{value}{the value of x to recode as NA}
}
\value{
a vector with values recoded to NA
}
\description{
Converts specified values in a vector to NA's.
}
\examples{
convert_to_na(x = c(1,2,3,4), value = 4)
}
