% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooking_recipes}
\alias{cooking_recipes}
\title{Cooking Recipes}
\format{
A data frame with 80 rows and 4 variables:
\describe{
  \item{name}{The name of the cooking recipe}
  \item{ingredients}{A space delimited set of pairs, where the first item in
  the pair is either an entire category (in which case, it is preceded by a
  negative sign) or the object id of a specific item, and where the second item
  in the pair is the number of objects required for the recipe}
  \item{yield_object_id}{The object id of the resulting item}
  \item{unlock_conditions}{A space delimited list of the conditions for unlocking
  the recipe. If not null, the first element is a letter that represents the type of
  event, with "f" for friendship, "l" for player level, and "s" for skill. If "l",
  then the next value is the level required. If "f" then the second element
  is the character to befriend and the third is the required friendship level.
  If "s," then the second element is the skill required and the third is the level
  required in this skill. If the unlock condition is "null" or "default", then it
  is either automatically unlocked at the start of the game, or it has a more
  complex storyline event in order to unlock it.}
  \item{liked_items}{object ID's of the liked items}
  \item{dislike_dialogue}{dialogue NPC uses when given a disliked item}
  \item{disliked_items}{object ID's of the disliked items}
  \item{hate_dialogue}{dialogue NPC uses when given a hated item}
  \item{hated_items}{object ID's of the hated items}
  \item{neutral_dialogue}{dialogue NPC uses when given a neutral item}
  \item{neutral_items}{object ID's of the neutral items}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Gift_taste_data}
}
\usage{
cooking_recipes
}
\description{
A dataset containing recipes that are made by cooking.
}
\keyword{datasets}
