% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crafting_recipes}
\alias{crafting_recipes}
\title{Crafting Recipes}
\format{
A data frame with 80 rows and 5 variables:
\describe{
  \item{name}{The name of the cooking recipe}
  \item{ingredients}{A space delimited set of pairs, where the first item in
  the pair is either an entire category (in which case, it is preceded by a
  negative sign) or the object id of a specific item, and where the second item
  in the pair is the number of objects required for the recipe}
  \item{yield_object_id}{The object id of the resulting item}
  \item{big_craftable}{Whether or not the object is a big craftable item, where
  possible values are "Big Craftable", "Ring" (if it is a ring), or "Other"}
  \item{unlock_conditions}{A space delimited list of the conditions for unlocking
  the recipe. If not null, the first element is a letter that represents the type of
  event, with "f" for friendship, "l" for player level, and "s" for skill. If "l",
  then the next value is the level required. If "f" then the second element
  is the character to befriend and the third is the required friendship level.
  If "s," then the second element is the skill required and the third is the level
  required in this skill. If the unlock condition "default", then it
  is automatically unlocked at the start of the game. Some events may have
  more complex unlock conditions described in text, rather than followin the
  above convention.}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Recipe_data}
}
\usage{
crafting_recipes
}
\description{
A dataset containing recipes that are made by crafting.
}
\keyword{datasets}
