% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crops}
\alias{crops}
\title{Crops Data}
\format{
A data frame with 42 rows and 22 variables:
\describe{
  \item{objectid}{unique id of the object, but note that this joins to the
  seed not the full-grown crop's id in other datasets}
  \item{days_in_stage_1_growth}{days in first stage of growth}
  \item{days_in_stage_2_growth}{days in second stage of growth}
  \item{days_in_stage_3_growth}{days in third stage of growth, and if NA then
  growth is completed prior to this stage}
  \item{days_in_stage_4_growth}{days in fourth stage of growth, and if NA then
  growth is completed prior to this stage}
  \item{days_in_stage_5_growth}{days in fifth stage of growth, and if NA then
  growth is completed prior to this stage}
  \item{growth_season_spring}{boolean representing if the crop grows in
  spring}
  \item{growth_season_summer}{boolean representing if the crop grows in
  summer}
  \item{growth_season_fall}{boolean representing if the crop grows in
  fall}
  \item{growth_season_winter}{boolean representing if the crop grows in
  winter}
  \item{index_in_sprite_sheet}{index of the sprite associated with this crop}
  \item{index_of_harvest}{objectid of the fully-grown crop, in constrast with
  `objectid` which is the unique id of the seed}
  \item{regrow_after_harvest}{-1 if the crop only grows once, and otherwise
  the number of days required to produce a new crop yield}
  \item{harvest_method}{1 if the item must be harvested with a scythe, else
  0}
  \item{chance_for_extra_harvest}{boolean representing if this plant allows
  probabilistic extra crop(s) at harvest}
  \item{min_extra_harvest}{the minimum number of extra crops produced, if an
  "extra harvest" event occurs}
  \item{max_extra_harvest}{the maximum number of extra crops produced, if an
  "extra harvest" event occurs}
  \item{max_harvest_increase_per_farming_level}{the maximum increase in crops
  per farming level, which adds to `max_extra_harvest`}
  \item{chance_for_extra_crops}{the probability of getting an extra harvest}
  \item{raised_seeds}{boolean representing if this crop grows on a raised
  item such as a pole subsequently and block your character from walking
  through this square}
  \item{tint_color}{boolean representing if ths item has special tints, e.g.
  certain flowers have various possible colors}
  \item{tint_color_value}{if `tint_color` is true, then sets of rgb values for
  each of the possible tints}
}

A data frame with 15 rows and 27 variables:
\describe{
  \item{name}{animal name}
  \item{days_to_produce}{days to produce item e.g. milk or egg}
  \item{age_when_mature}{days to achieve maturity}
  \item{default_produce_index}{object id of regular quality item produced}
  \item{deluxe_produce_index}{object id of deluxe quality item produced}
  \item{sound}{sound the animal makes}
  \item{bounding_box_front_x}{x coordinate of the front-view bounding box}
  \item{bounding_box_front_y}{y coordinate of the front-view bounding box}
  \item{bounding_box_front_width}{width of front-view bounding box}
  \item{bounding_box_front_height}{height of front-view bounding box}
  \item{bounding_box_side_x}{x coordinate of the side-view bounding box}
  \item{bounding_box_side_y}{y coordinate of the side-view bounding box}
  \item{bounding_box_side_width}{width of side-view bounding box}
  \item{bounding_box_side_height}{height of side-view bounding box}
  \item{harvest_type}{method of harvesting}
  \item{change_texture_when_item_ready}{does the animal's coat change texture
  when it is ready for harvest?}
  \item{building_type}{type of building the animal lives in}
  \item{sprite_size_front_width}{front-view width of the sprite}
  \item{sprite_size_front_height}{front-view height of the sprite}
  \item{sprite_size_side_width}{side-view width of the sprite}
  \item{sprite_size_side_height}{side-view height of the sprite}
  \item{fullness_drain}{rate of decay after feeding}
  \item{happiness_drain}{rate of decay after saying hello}
  \item{meat_index}{object id of the meat if the animal is butchered}
  \item{sell_price}{sell price of the meat if the animal is butchered}
  \item{english_display_type}{English-language name for the animal}
  \item{english_display_building}{English-langauge name for the building}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Crop_data}

Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Animal_data}
}
\usage{
crops

crops
}
\description{
A dataset containing information about crops in Stardew Valley.

A dataset containing information about animals in Stardew Valley.
}
\keyword{datasets}
