% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crops_object_information}
\alias{crops_object_information}
\title{Crops Object Information}
\format{
A data frame with 42 rows and 22 variables:
\describe{
  \item{objectid}{unique id of the crop, which joins to `crops` columns
  `index_of_harvest`}
  \item{name}{The name of the crop}
  \item{price}{The sell price of the crop (note that this is sell price, not purchase price)}
  \item{edibility}{Value used to determine health/energy regained by eating the item.
  The energy gained is 2.5 x `edibility`, and the health gained is 1.125 x `edibility`.
  Note that the amount of health shown in the tooltip is only 1 x `edibility` but
  the actual amount gained is 1.125 x `edibility`.}
  \item{type}{The object type}
  \item{category}{The category of the crop}
  \item{english_name}{The name of the object in English, and a duplicate of
  `name` (may update in the future to include other language names)}
  \item{description}{A description of the crop}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Object_data}
}
\usage{
crops_object_information
}
\description{
A dataset containing more detailed information about crops in Stardew Valley.
If you want information about the seeds and growing patterns which produce
these crops, you may way to join to the `crops` table on `object_id` =
`index_of_harvest`.
}
\keyword{datasets}
