% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foods_and_drinks}
\alias{foods_and_drinks}
\title{Foods And Drinks}
\format{
A data frame with 103 rows and 21 variables:
\describe{
  \item{object_id}{Object id of the food or drink item}
  \item{name}{Name of the food or drink item}
  \item{price}{Price of the food or drink item}
  \item{edibility}{Value used to determine health/energy regained by eating the item.
  The energy gained is 2.5 x `edibility`, and the health gained is 1.125 x `edibility`.
  Note that the amount of health shown in the tooltip is only 1 x `edibility` but
  the actual amount gained is 1.125 x `edibility`.}
  \item{category}{The category of the food or drink item}
  \item{type}{Type of object food or drink item}
  \item{english_name}{The name of the object in English, and a duplicate of
  `name` (may update in the future to include other language names)}
  \item{description}{Description of the food or drink item}
  \item{food_or_drink}{Is the item food or drink?}
  \item{buff_farming}{}
  \item{buff_fishing}{}
  \item{buff_mining}{}
  \item{buff_digging_unimplemented}{}
  \item{buff_luck}{}
  \item{buff_foraging}{}
  \item{buff_crafting}{}
  \item{buff_max_energy}{}
  \item{buff_magnetism}{}
  \item{buff_speed}{}
  \item{buff_defense}{}
  \item{buff_duration}{}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Object_data}
}
\usage{
foods_and_drinks
}
\description{
A dataset containing information about foods and drinks in Stardew Valley.
Foods and drinks can be purchased and crafted.
}
\keyword{datasets}
