% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{furniture}
\alias{furniture}
\title{Furniture}
\format{
A data frame with 347 rows and 9 variables:
\describe{
  \item{objectid}{Unique id of the furniture item}
  \item{name}{The name of the piece of furniture}
  \item{source_rectangle_width}{The width of the rectangle of the item}
  \item{source_rectangle_width}{The height of the rectangle of the item}
  \item{bounding_box_width}{The width of the object's bounding box}
  \item{bounding_box_height}{The height of the object's bounding box}
  \item{rotations}{The number of possible rotations of the object, where 1 =
  only one view, 2 = two rotations (180 degrees), and 4 = four rotations (90
  degrees)}
  \item{price}{The price of the furniture}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Furniture_data}
}
\usage{
furniture
}
\description{
A dataset containing information about furniture in Stardew Valley.
}
\keyword{datasets}
