% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geodes}
\alias{geodes}
\title{Geodes}
\format{
A data frame with 4 rows and 7 variables:
\describe{
  \item{object_id}{Object id of the geode}
  \item{name}{Name of the geode}
  \item{price}{Price of the geode}
  \item{edibility}{Edibility typically represents a variable used to
  calculated the number of health and energy points regained by eating
  the item, but all geodes are inedible}
  \item{type}{Type of object (always Basic)}
  \item{english_name}{The name of the object in English, and a duplicate of
  `name` (may update in the future to include other language names)}
  \item{description}{Description of the geode}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Object_data}
}
\usage{
geodes
}
\description{
A dataset containing information about the geodes in Stardew Valley. Geodes
can be broken open by the blacksmith and may reveal certain items inside,
based upon their probabilities of appearing.
}
\keyword{datasets}
