% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hats}
\alias{hats}
\title{Hats}
\format{
A data frame with 94 rows and 6 variables:
\describe{
  \item{object_id}{Unique id of the hat}
  \item{name}{The name of the hat}
  \item{description}{Description of the hat}
  \item{skip_hair_draw}{If true, then the hat does not change the player's
  hairstyle, but if false, then it may}
  \item{ignore_hairstyle_offset}{If true, then the hat occupies a fixed position,
  but if false, then it adjsut to the line of the player's hairstyle}
  \item{prismatic}{Is the hat a prismatic item?}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Hat_data}
}
\usage{
hats
}
\description{
A dataset containing information about hats in Stardew Valley. Hats can be
purchased from the hat shop owned by the mouse, and new hats are unlocked
when certain in-game achievements are completed.
}
\keyword{datasets}
