% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations}
\format{
A data frame with 23 rows and 10 variables:
\describe{
  \item{name}{Name of the location}
  \item{spring_foraging}{Object id's and probabilities of items which can be foraged in this area
  during spring}
  \item{summer_foraging}{Object id's and probabilities of items which can be foraged in this area
  during summer}
  \item{fall_foraging}{Object id's and probabilities of items which can be foraged in this area
  during fall}
  \item{winter_foraging}{Object id's and probabilities of items which can be foraged in this area
  during winter}
  \item{spring_fishing}{Object id's and probabilities of fish which can be caught in this area
  during spring}
  \item{summer_fishing}{Object id's and probabilities of fish which can be caught in this area
  during summer}
  \item{fall_fishing}{Object id's of and probabilities fish which can be caught in this area
  during fall}
  \item{winter_fishing}{Object id's and probabilities of fish which can be caught in this area
  during winter}
  \item{artifact_data}{Object id's and probabilities of artifacts which can be found in this area}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Hat_data}
}
\usage{
locations
}
\description{
A dataset containing information about locations in Stardew Valley.
Locations have different foraging and fishing options depending on the time
of the year. Each datapoint in this dataset contains a space delimited list
of pairs. For foraging, this alternates the object id of the item with the probability it
appears (e.g. '384 .08 589 .09' means that object 384 has an 8% probability
of appearing and object 589 has a 9% probability of appearing). For
fish, this alternates the object id with the location id. Location id is -1 for
any pond or river, 0 for river.
}
\keyword{datasets}
