% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{npc_gift_tastes}
\alias{npc_gift_tastes}
\title{NPC Gift Tastes}
\format{
A data frame with 738 rows and 7 variables:
\describe{
  \item{npc_name}{the NPC name}
  \item{dialogue_type}{the type of dialogue (liked, loved, etc.)}
  \item{dialogue}{the dialogue the npc will say when receiving the gift}
  \item{item_type}{the item type (liked, loved, etc.)}
  \item{object_id}{the unique id of the object, which can join to `objects`}
  \item{object_name}{the name of the object, if it is an object}
  \item{category_id}{the unique id of the category, which can join to `objects`}
  \item{category_name}{the name of the category, if it is a category}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewvalleywiki.com/Modding:Gift_taste_data}
}
\usage{
npc_gift_tastes
}
\description{
A dataset containing NPC gift tastes that are specific to that NPC,
not including universal tastes. For universal gift tastes, see
`universal_gift_tastes`, and for the two sets pre-combined, see
`complete_npc_gift_tastes`.
}
\keyword{datasets}
