% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdew-package.R
\docType{package}
\name{rdew-package}
\alias{rdew}
\alias{rdew-package}
\title{rdew: rdew}
\description{
Holds data and functions related to the game Stardew Valley
}
\author{
\strong{Maintainer}: Afton Coombs \email{aftoncoombs@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Kyle Harris \email{koderkow@gmail.com}
}

}
\keyword{internal}
