% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rings}
\alias{rings}
\title{Rings}
\format{
A data frame with 20 rows and 7 variables:
\describe{
  \item{object_id}{Object id of the ring}
  \item{name}{Name of the ring}
  \item{price}{Price of the ring}
  \item{edibility}{Edibility typically represents a variable used to
  calculated the number of health and energy points regained by eating
  the item, but all rings are inedible}
  \item{type}{Type of object (always Ring)}
  \item{english_name}{The name of the object in English, and a duplicate of
  `name` (may update in the future to include other language names)}
  \item{description}{Description of the ring}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Object_data}
}
\usage{
rings
}
\description{
A dataset containing information about rings in Stardew Valley. Rings,
with the exception of the wedding ring, provide a player with buffs and
special abilities.
}
\keyword{datasets}
