% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_splitter.R
\name{split_column}
\alias{split_column}
\title{Split Ingredients}
\usage{
split_column(.data, cols, ...)
}
\arguments{
\item{.data}{a data.frame or data.frame-like object}
}
\value{
a long-form dataset with columns for required object and/or quantity of required object
}
\description{
Splits the column `required_ingredients` into required object and quantity of
required object, for all required objects, and converts the dataset to long format
with columns for required object and quantity of required object.
}
