% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ingredients.R
\name{split_ingredients}
\alias{split_ingredients}
\title{Split Ingredients}
\usage{
split_ingredients(x)
}
\arguments{
\item{x}{a data.frame or data.frame-like object containing a column
`required_ingredients` which is itself a space-delimited set of pairs}
}
\value{
a long-form dataset with columns for required object and quantity of
required object
}
\description{
Splits the column `required_ingredients` into required object and quantity of
required object, for all required objects, and converts the dataset to long format
with columns for required object and quantity of required object.
}
\examples{
split_ingredients(x = crafting_recipes)
}
