% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{universal_gift_tastes}
\alias{universal_gift_tastes}
\title{Universal Gift Tastes}
\format{
A data frame with 88 rows and 4 variables:
\describe{
  \item{universal_type}{type of the object (e.g. liked or loved)}
  \item{object_id}{the unique identifier of the object, which can be joined
  to `objects`, where a negative value represents a category and not a singular object}
  \item{object_name}{the name of the object if it is an object}
  \item{category_name}{the name of the object if it is a category}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game. Descriptions of the variables summarized from
\url{https://stardewcommunitywiki.com/Modding:Gift_taste_data}
}
\usage{
universal_gift_tastes
}
\description{
A dataset containing universal gift tastes, which apply for every NPC
unless that NPC specifies in `npc_gift_tastes` a different preference for that
item, which overrules `universal_gift_tastes`.
}
\keyword{datasets}
