% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{building_blueprints}
\alias{building_blueprints}
\title{Building Blueprints Data}
\format{
A data frame with 25 rows and 20 variables:
\describe{
  \item{name}{the name of the building}
  \item{items_required}{a space-delimited string containing the object
  ids of the items required to construct the building, if any}
  \item{tiles_width}{width of the building in game tiles}
  \item{tiles_height}{height of the building in game tiles}
  \item{human_door_x_coord}{x-coordinate of the door that players enter
  through}
  \item{human_door_y_coord}{y-coordinate of the door that players enter
  through}
  \item{animal_door_x_coord}{x-coordinate of the door that animals enter
  through}
  \item{animal_door_y_coord}{y-coordinate of the door that animals enter
  through}
  \item{map_to_warp_to}{the map image the game uses for the interior of
  the building, if any}
  \item{english_name}{the name of the object in English, and a duplicate of
  `name` (may update in the future to include other language names)}
  \item{description}{a description of the building}
  \item{blueprint_type}{whether a building is an initial building or
  an upgraded version of a previous building}
  \item{name_of_building_to_upgrade}{if building type is 'upgrade', then the
  name of the building to upgrade}
  \item{source_rectangle_for_menu_view_1}{x-coordinate of the menu view
  rectangle}
  \item{source_rectangle_for_menu_view_y}{y-coordinate of the menu view
  rectangle}
  \item{max_occupants}{maximum number of occupants}
  \item{action_behavior}{action the building provides, if any}
  \item{names_of_building_location}{where the building is located}
  \item{price}{cost of having the building constructed}
  \item{magical}{is the building magical?}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game.
Descriptions of the variables summarized from \url{https://stardewvalleywiki.com/Modding:Big_craftables_data}
}
\usage{
building_blueprints
}
\description{
A dataset containing information about blueprints for buildings you can
purchase or construct.
}
\keyword{datasets}
