% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundles}
\alias{bundles}
\title{Bundles Data}
\format{
A data frame with 31 rows and 11 variables:
\describe{
  \item{room_id_and_sprite_index}{Unique id of the room and sprite, which
  when split by / will return the name of the Community Center room and the
  id of the sprite for bundle (each room has multiple bundles)}
  \item{bundle_name}{the name of the bundle}
  \item{reward_type_of_object}{The type of the reward earned by delivering
  the bundle, where the type may be O for "object," BO for "big object,"
  F for "furniture," H  for "hat", C for "clothing", or R for "ring". A bundle
  reward may include more than one type.}
  \item{reward_object_id}{id of the object given as a reward
  for completing the bundle}
  \item{reward_number_of_object_given}{The count of the object given as a reward}
  \item{required_object_id}{Space-delimited id's for the objects required
  to complete the bundle}
  \item{required_object_is_gold}{boolean representing if the object required
  to complete the quest is gold}
  \item{required_number_of_objects}{Required number of each object, if
  submitted to complete the bundle}
  \item{required_min_quality_of_objects}{Required quality of the object for the
  bundle}
  \item{color_index}{Color index of the bundle color}
  \item{min_number_of_items}{a description of the building}
}
}
\source{
Data loaded from \url{https://github.com/aftonsteps/stardewdata} which
in turn depends on \url{https://github.com/LeonBlade/xnbcli} for extracting the data
from the game.
Descriptions of the variables summarized from \url{https://stardewvalleywiki.com/Modding:Bundles}
}
\usage{
bundles
}
\description{
A dataset containing information about the bundles delivered to the Community
Center.
}
\keyword{datasets}
